Pushed colors to shader, reduced texture size
Background/foreground colors are now set in the shader.
Because of this, can reduce the size of the font bitmaps to 1 channel, reducing memory usage.
Added new state; need to have pixel unpack align be 1, since there are no garuntees with single channel textures.
Also changed shaders to be embedded.