Fixed typo that caused bad index correction
Instead of checking if string[i] == env::CONTROL_CHAR, it cheked if `i` itself was. This resulted in bad index offsets, which resulted in the control block offsets being chopped off and replaced, causing bad offsets to be read as locations for the text mods, causing segmentation faults from dereferencing null pointers. This was a fun morning~