Reduced time complexity of look behinds while rendering
Normally, you'd have to run from the start to the render-start, recording all the color settings.
Now however, we read back on a per-line basis. If the Lookbehind reports that there was an attempt to change its background and foreground after checking the line, then we know that we have the most recent color settings.

The way I did it was perhaps a little bit of a hack, but I personally like it.